Move udp-client to its own crate
