+++ tpu-client/src/lib.rs
@@ -6,7 +6,6 @@ pub mod quic_client;
 pub mod tpu_client;
 pub mod tpu_connection;
 pub mod tpu_connection_cache;
-pub mod udp_client;
 
 #[macro_use]
 extern crate solana_metrics;
