+++ tpu-client/src/nonblocking/mod.rs
@@ -1,4 +1,3 @@
 pub mod quic_client;
 pub mod tpu_client;
 pub mod tpu_connection;
-pub mod udp_client;
