+++ udp-client/Cargo.toml
@@ -10,6 +10,10 @@ documentation = "https://docs.rs/solana-udp-client"
 edition = "2021"
 
 [dependencies]
+async-trait = "0.1.57"
 solana-net-utils = { path = "../net-utils", version = "=1.15.0" }
+solana-sdk = { path = "../sdk", version = "=1.15.0" }
+solana-streamer = { path = "../streamer", version = "=1.15.0" }
 solana-tpu-client = { path = "../tpu-client", version = "=1.15.0" }
 thiserror = "1.0"
+tokio = { version = "1", features = ["full"] }
